[configs] Classify fmradio loopback source as player. JB#41613
Instead of classifying source output of module-loopback as internal
(meaning policy enforcement doesn't touch the source output at all)
classify the source output as player. This way when we have multiple
sources to choose from policy enforcement can connect the source output
to correct source.

Now when for example USB-audio device with recording capabilities is
connected when fmradio is enabled, module-loopback connects to the
first source it sees which may be the wrong one. With this change
policy enforcement has pointers to select the right source.

Signed-off-by: Juho Hämäläinen <juho.hamalainen@jolla.com>